restructure lambda
