+++ ops/aws/canary/lambda/main.go
@@ -5,30 +5,14 @@ import (
 	"fmt"
 	"github.com/aws/aws-lambda-go/lambda"
 	"github.com/filecoin-project/bacalhau/cmd/bacalhau"
+	"github.com/filecoin-project/bacalhau/ops/aws/canary/scenarios"
+	"github.com/filecoin-project/bacalhau/pkg/publicapi"
 	"github.com/filecoin-project/bacalhau/pkg/system"
 )
 
-type Event struct {
-	Action string `json:"action"`
-}
-
-func hello(event Event) (string, error) {
-	fmt.Printf("Received event: %+v", event)
-
-	jobs, err := bacalhau.GetAPIClient().List(context.Background())
-	if err != nil {
-		return "", err
-	}
-
-	count := 0
-	for _, j := range jobs {
-		fmt.Printf("Job: %s\n", j.ID)
-		count++
-		if count > 10 {
-			break
-		}
-	}
-	return "Done λ!", nil
+var client *publicapi.APIClient
+var testcasesMap = map[string]Handler{
+	"list": scenarios.List,
 }
 
 func init() {
@@ -36,8 +20,21 @@ func init() {
 	if err != nil {
 		panic(err)
 	}
+	client = bacalhau.GetAPIClient()
+}
+
+func route(ctx context.Context, event Event) error {
+	handler, ok := testcasesMap[event.Action]
+	if !ok {
+		return fmt.Errorf("no handler found for action: %s", event.Action)
+	}
+	err := handler(ctx, client)
+	if err != nil {
+		return err
+	}
+	return nil
 }
 
 func main() {
-	lambda.Start(hello)
+	lambda.Start(route)
 }
